zombies only chase when in sight

+++ Source/TanksVSZombies/Tank.cpp
@@ -161,7 +161,7 @@ void ATank::SetupPlayerInputComponent(UInputComponent* PlayerInputComponent)
 	InputComponent->BindAction(TEXT("ShootSecondary"), IE_Released, this, &ATank::ShootSecondaryReleased);
 }
 
-void ATank::DamageHealth(float damage)
+void ATank::DamageHealth(int damage)
 {
 	//TODO
 }
